+++ domain/src/main/java/fr/insee/rem/domain/ports/api/PartitionServicePort.java
@@ -14,4 +14,5 @@ public interface PartitionServicePort {
 
     List<PartitionDto> getAllPartitions();
 
+    void emptyPartitionById(Long partitionId);
 }