+++ domain/src/main/java/fr/insee/rem/domain/ports/api/SurveyUnitServicePort.java
@@ -30,4 +30,5 @@ public interface SurveyUnitServicePort {
 
     SurveyUnitDto updateSurveyUnit(SurveyUnitDto surveyUnit);
 
+    long countSurveyUnitsByPartition(Long partitionId);
 }