+++ domain/src/main/java/fr/insee/rem/domain/ports/spi/PartitionSurveyUnitLinkPersistencePort.java
@@ -20,4 +20,7 @@ public interface PartitionSurveyUnitLinkPersistencePort {
 
     List<SuIdMappingRecord> findSurveyUnitIdsMappingTableByPartitionId(Long partitionId);
 
+    long countByPartitionId(Long partitionId);
+
+    void removeSurveyUnitsFromPartition(Long partitionId);
 }