+++ infrastructure/src/main/java/fr/insee/rem/infrastructure/entity/AdditionalInformation.java
@@ -4,17 +4,9 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import lombok.Data;
 
-import java.io.Serializable;
-
 @Data
 @JsonInclude(Include.NON_EMPTY)
-public class AdditionalInformation implements Serializable {
-
-    /**
-     *
-     */
-    private static final long serialVersionUID = -1278478308717480850L;
-
+public class AdditionalInformation {
     private String key;
     private String value;
 }