+++ infrastructure/src/main/java/fr/insee/rem/infrastructure/entity/Address.java
@@ -4,15 +4,9 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import lombok.Data;
 
-import java.io.Serializable;
-
 @Data
 @JsonInclude(Include.NON_EMPTY)
-public class Address implements Serializable {
-    /**
-     *
-     */
-    private static final long serialVersionUID = -5012898811152555612L;
+public class Address {
     private String streetNumber;
     private String repetitionIndex;
     private String streetType;