+++ infrastructure/src/main/java/fr/insee/rem/infrastructure/entity/SurveyUnitData.java
@@ -4,17 +4,11 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import lombok.Data;
 
-import java.io.Serializable;
 import java.util.List;
 
 @Data
 @JsonInclude(Include.NON_EMPTY)
-public class SurveyUnitData implements Serializable {
-
-    /**
-     *
-     */
-    private static final long serialVersionUID = 811623138921050269L;
+public class SurveyUnitData {
 
     private Address address;