+++ infrastructure/src/main/java/fr/insee/rem/infrastructure/entity/SurveyUnitEntity.java
@@ -1,6 +1,7 @@
 package fr.insee.rem.infrastructure.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.databind.JsonNode;
 import fr.insee.rem.domain.dtos.Context;
 import jakarta.persistence.*;
 import lombok.Data;
@@ -9,19 +10,13 @@ import lombok.ToString;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 
-import java.io.Serializable;
 import java.util.HashSet;
 import java.util.Set;
 
 @Entity
 @Data
 @Table(name = "survey_unit")
-public class SurveyUnitEntity implements Serializable {
-
-    /**
-     *
-     */
-    private static final long serialVersionUID = 5458959377889736946L;
+public class SurveyUnitEntity {
 
     @Id
     @GeneratedValue(generator = "seq_survey_unit", strategy = GenerationType.SEQUENCE)
@@ -38,6 +33,10 @@ public class SurveyUnitEntity implements Serializable {
     @Column(columnDefinition = "jsonb", name = "data")
     @Basic(fetch = FetchType.EAGER)
     private SurveyUnitData surveyUnitData;
+
+    @JdbcTypeCode(SqlTypes.JSON)
+    @Column(columnDefinition = "json", name = "externals")
+    private JsonNode externals;
 
     @OneToMany(mappedBy = "surveyUnit", orphanRemoval = false, cascade = CascadeType.REMOVE)
     @JsonIgnore